Always generate :pages up-front, not in the page-list transform
This simplifies page list generation. The call to the children
function is only done at one point now, when pages are first
created.

The :requires-split? key is no longer needed.

This code only sets :pages for the first item in :page-list-specs.
More work would need to be done to support multiple :pages.